Added mesh structure and simple generators/loaders

+++ src/lib/CG.h
@@ -14,7 +14,7 @@ namespace CG
 		template<template<typename> class M, typename T>
 		typename std::enable_if<is_matrix<M, T>::value, DynamicVector<T> >::type
 			cg(const M<T>& a, const DynamicVector<T>& b, const DynamicVector<T>& x0,
-				uint32 maxIter = 1024, double eps = 10e-6, uint32* it_stat = nullptr);
+				size_t maxIter = 1024, double eps = 10e-6, size_t* it_stat = nullptr);
 	}
 }
